Add storage engine support to Predictor
