+++ README.rst
@@ -25,28 +25,41 @@ Features
 
 Usage
 --------
-```
-from triage.model_trainers import SimpleModelTrainer
-from triage.predictors import Predictor
 
-db_engine = ...
-s3_conn = ...
-project_path = 'econ-dev/inspections'
+.. code-block::
+    from triage.storage import S3ModelStorageEngine
+    from triage.model_trainers import SimpleModelTrainer
+    from triage.predictors import Predictor
 
-trainer = SimpleModelTrainer(
-    training_set_path=...,
-    training_metadata_path=...,
-    model_config=...,
-    project_path=project_path,
-    s3_conn=s3_conn,
-    db_engine=db_engine
-)
-model_ids = trainer.train_models()
+    db_engine = ...
+    s3_conn = ...
+    project_path = 'econ-dev/inspections'
+    model_storage_engine = S3ModelStorageEngine(s3_conn, project_path)
 
-predictor = Predictor(project_path, s3_conn, engine)
-for model_id in model_ids:
-    predictions = predictor.predict(model_id, test_matrix_path=..., test_metadata_path=...)
-```
+    trainer = SimpleModelTrainer(
+        project_path=project_path,
+        model_storage_engine,
+        db_engine=db_engine
+    )
+    predictor = Predictor(
+        project_path=project_path,
+        model_storage_engine=model_storage_engine,
+        db_engine=db_engine
+    )
+
+    model_ids = trainer.train_models(
+        training_set_path=...,
+        training_metadata_path=...,
+        model_config=...
+    )
+
+    for model_id in model_ids:
+        predictions = predictor.predict(
+            model_id,
+            test_matrix_path=...,
+            test_metadata_path=...
+        )
+
 
 Credits
 ---------
