+++ triage/storage.py
@@ -1,0 +1,68 @@
+from .utils import upload_object_to_key, key_exists, model_cache_key, download_object
+import os
+import pickle
+
+
+class ModelStorageEngine(object):
+    def __init__(self, project_path):
+        self.project_path = project_path
+
+    def get_store(self, model_hash):
+        pass
+
+
+class Store(object):
+    def __init__(self, path):
+        self.path = path
+
+    def exists(self):
+        pass
+
+    def load(self):
+        pass
+
+
+class S3Store(Store):
+    def exists(self):
+        key_exists(self.path)
+
+    def write(self, obj):
+        upload_object_to_key(obj, self.path)
+
+    def load(self):
+        return download_object(self.path)
+
+
+class FSStore(Store):
+    def exists(self):
+        return os.path.isfile(self.path)
+
+    def write(self, obj):
+        with open(self.path, 'w+b') as f:
+            pickle.dump(obj, f)
+
+    def load(self):
+        with open(self.path, 'rb') as f:
+            return pickle.load(f)
+
+
+class S3ModelStorageEngine(ModelStorageEngine):
+    def __init__(self, s3_conn, *args, **kwargs):
+        super(S3ModelStorageEngine, self).__init__(*args, **kwargs)
+        self.s3_conn = s3_conn
+
+    def get_store(self, model_hash):
+        return S3Store(model_cache_key(
+            self.project_path,
+            model_hash,
+            self.s3_conn
+        ))
+
+
+class FSModelStorageEngine(ModelStorageEngine):
+    def get_store(self, model_hash):
+        return FSStore('/'.join([
+            self.project_path,
+            'trained_models',
+            model_hash
+        ]))
